+++ CHANGES.rst
@@ -1,6 +1,8 @@
 Unreleased
 ==========
 
+- [IMPROVED] Accept slurm-style host lists
+- [NEW] Add status command for setting status
 - [IMPROVED] Move tox versions to python 3.6, 3.7, 3.8
 
 0.2.1 (2018-02-19)